r2086/sofa-dev : FIX: warnings on Windows
 //partial sync (9/11 changes)//

+++ modules/sofa/component/constraint/BilateralInteractionConstraint.inl
@@ -83,7 +83,7 @@ void BilateralInteractionConstraint<DataTypes>::getConstraintId(long* id, unsign
 {
     if (!yetIntegrated)
     {
-        id[offset++] = -cid;
+        id[offset++] = -(int)cid;
 
         yetIntegrated = true;
     }
